Move distance functions into data init
Add tests into test_data
