+++ mousestyles/data/tests/test_data.py
@@ -114,3 +114,42 @@ def test_start_time_end_time_loader():
     times = data.load_start_time_end_time(0, 0, 0)
     assert len(times) == 2
     assert times[1] - times[0] > 0
+
+
+def test_pos_day():
+    assert np.all(data.distances(1, 2, 3) >= 0)
+
+
+def test_pos_bymouse():
+    assert np.all(data.distances_bymouse(1, 2, 3) >= 0)
+
+
+def test_pos_bystrain():
+    assert np.all(data.distances_bystrain(1, 2, 3) >= 0)
+
+
+def test_array_day():
+    assert type(data.distances(1, 2, 3)) is np.ndarray
+
+
+def test_array_bymouse():
+    assert type(data.distances_bymouse(1, 2)) is np.ndarray
+
+
+def test_array_bystrain():
+    assert type(data.distances_bystrain(1)) is np.ndarray
+
+
+def max_speed():
+    # Max speed of a mouse should be less than 40 km/h
+    assert max(data.distances(0, 0, 0, step=50) * 3.6 / 100) < 40
+
+
+def max_speed_bymouse():
+    # Max speed of a mouse should be less than 40 km/h
+    assert max(data.distances_bymouse(0, 0, step=50) * 3.6 / 100) < 40
+
+
+def max_speed_bystrain():
+    # Max speed of a mouse should be less than 40 km/h
+    assert max(data.distances_bystrain(0, step=50) * 3.6 / 100) < 40
